+++ comp2013/src/main/java/comp2013/View/SnakeFood.java
@@ -121,7 +121,7 @@ public class SnakeFood
 	 * @param h2 second image height
 	 * @return Returns true if the image are intersecting, false otherwise.
 	 */
-	private boolean areImagesIntersecting(double x1, double y1, double w1, double h1,
+	public static boolean areImagesIntersecting(double x1, double y1, double w1, double h1,
 										  double x2, double y2, double w2, double h2) {
 		return (x1 < x2 + w2 && x1 + w1 > x2 && y1 < y2 + h2 && y1 + h1 > y2);
 	}